requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.3
redis==5.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==0.20.0
flask-caching==2.3.0
faker==28.4.1